backend/requirements.txt
# Core framework
fastapi>=0.110
uvicorn[standard]>=0.27

# Database & ORM
sqlalchemy>=2.0
psycopg2-binary>=2.9
alembic>=1.13

# Environment & settings
python-dotenv>=1.0
pydantic>=2.6
pydantic-settings>=2.2
pydantic-extra-types>=2.5

# Async & networking (used by FastAPI)
anyio>=4.0
httpx>=0.27
httptools>=0.6
websockets>=12.0
watchfiles>=0.21

# Templates & CLI (Alembic / FastAPI tooling)
jinja2>=3.1
mako>=1.3
typer>=0.9
rich>=13.7
rich-toolkit>=0.11

# Utilities
python-multipart>=0.0.9
email-validator>=2.1
pyyaml>=6.0
certifi
urllib3
idna
supabase
celery
redis
google-generativeai
python-jose[cryptography]
